+++ edgy/core/utils/db.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
     from edgy.core.connection.database import Database
 
 
-def check_db_connection(db: "Database") -> None:
+def check_db_connection(db: "Database", stacklevel: int = 3) -> None:
     if not db.is_connected:
         # with force_rollback the effects are even worse, so fail
         if db.force_rollback:
@@ -17,7 +17,7 @@ def check_db_connection(db: "Database") -> None:
         warnings.warn(
             "Database not connected. Executing operation is inperformant.",
             UserWarning,
-            stacklevel=2,
+            stacklevel=stacklevel,
         )